+++ exercise-akka-persistence-basic/src/test/scala/testing/tagobjects/ExerciseTest.scala
@@ -1,0 +1,5 @@
+package testing.tagobjects
+
+import org.scalatest.Tag
+
+object ExerciseTest extends Tag("example.testing.tags.ExerciseTest")